Introduce SoftwareDrawingContext and partially refactor things from Gfx
