+++ src/OpenLoco/CMakeLists.txt
@@ -63,6 +63,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Date.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/DrawSprite.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/FPSCounter.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/SoftwareDrawingContext.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/SoftwareDrawingEngine.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Economy/Economy.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/EditorController.cpp"
@@ -308,6 +309,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Date.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/DrawSprite.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/FPSCounter.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/SoftwareDrawingContext.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Drawing/SoftwareDrawingEngine.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Economy/Currency.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Economy/Economy.h"
